+++ tests/Container/HArrayAppendTest.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace Haystack\Tests\Container;
 
 use Haystack\HArray;
@@ -12,7 +14,7 @@ class HArrayAppendTest extends TestCase
     /** @var  HArray */
     private $arrDict;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->arrList = new HArray(['apple', 'bobble', 'cobble', 'dobble']);
         $this->arrDict = new HArray(['a' => 'apple', 'b' => 'bobble', 'c' => 'cobble', 'd' => 'dobble']);
@@ -25,7 +27,7 @@ class HArrayAppendTest extends TestCase
      * @param mixed $newThing
      * @param HArray $expected
      */
-    public function testAppendStringInArray($type, $newThing, $expected)
+    public function testAppendStringInArray(string $type, $newThing, HArray $expected): void
     {
         if ('list' === $type) {
             $newArray = $this->arrList->append($newThing);
@@ -36,7 +38,7 @@ class HArrayAppendTest extends TestCase
         $this->assertEquals($expected, $newArray);
     }
 
-    public function appendProvider()
+    public function appendProvider(): array
     {
         $dateTime = new \DateTime();