+++ tests/Container/HArrayRemoveTest.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace Haystack\Tests\Container;
 
 use Haystack\HArray;
@@ -8,10 +10,11 @@ class HArrayRemoveTest extends TestCase
 {
     /** @var  HArray */
     private $arrList;
+
     /** @var  HArray */
     private $arrDict;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->arrList = new HArray(['apple', 'bobble', 'cobble', 'dobble']);
         $this->arrDict = new HArray(['a' => 'apple', 'b' => 'bobble', 'c' => 'cobble', 'd' => 'dobble']);
@@ -24,7 +27,7 @@ class HArrayRemoveTest extends TestCase
      * @param string $value
      * @param HArray $expected
      */
-    public function testArrayTypeRemove($type, $value, $expected)
+    public function testArrayTypeRemove(string $type, string $value, HArray $expected): void
     {
         if ('list' === $type) {
             $newArr = $this->arrList->remove($value);
@@ -36,7 +39,7 @@ class HArrayRemoveTest extends TestCase
 
     }
 
-    public function arrayRemoveProvider()
+    public function arrayRemoveProvider(): array
     {
         return [
             'List: Basic list' => ['list', 'bobble', new HArray(['apple', 'cobble', 'dobble'])],
@@ -46,7 +49,7 @@ class HArrayRemoveTest extends TestCase
         ];
     }
 
-    public function testArrayTypeRemoveObject()
+    public function testArrayTypeRemoveObject(): void
     {
         $timestamp = new \DateTime();