+++ tests/Container/HStringRemoveTest.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace Haystack\Tests\Container;
 
 use Haystack\HString;
@@ -12,13 +14,13 @@ class HStringRemoveTest extends TestCase
     /** @var HString */
     protected $utf8String;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->aString = new HString('foobar');
         $this->utf8String = new HString('ɹɐqooɟ');
     }
 
-    public function testTypesOfStringRemove()
+    public function testTypesOfStringRemove(): void
     {
         $newString = $this->aString->remove('o');
         $this->assertEquals(new HString('fobar'), $newString);